Use system wxWidgets on Linux if custom build is not available.
Not particularly fond of the naming of "WX_CONFIG_CUSTOM" but have no ohter
ideas. Thought about appending the name of the variable that points to the
custom build directory but that does not really work now when it is
"WXWIDGETS". Changing the name of this variable would disrupt build
environments people have already set up so leave it as is for now.